% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Myfunction.R
\name{MyFunction}
\alias{MyFunction}
\title{MyFunc}
\usage{
MyFunction(a, b)
}
\arguments{
\item{a}{is the first positive input}

\item{b}{is the second positive input}
}
\value{
the product of the two numbers
}
\description{
My function takes two positive inputs and returns
their products
}
